Fix syntax struct in RPCE header.
The struct is described in C706 12.6.3 as following:

typedef   struct   {
             uuid_t   if_uuid;
             u_int32   if_version;
          } p_syntax_id_t;

For abstract syntax, if_uuid is set to the interface UUID, and if_version is
set to the interface version.
For transfer syntax, these are set to the UUID and version created for the
data representation. The major version is encoded in the 16 least significant
bits of if_version and the minor version in the 16 most significant bits.

So the major/minor version should not be two uint16 fields.