Merged PR 1236: [CaseFix] Fix the random file names generation issue which may occur after the previous
Fix the random file names generation issue which may occur after the previous fix.
`Random` objects will use the same seed if they are created in a short interval, and `ThreadStatic` attribute will not prevent the issue if the instances of different threads are created at the same time and the case will involve them during its execution.