[MSRDPEGFX]: add ExpectChannelClosed adapter method
The EvictCachEntry test for a non existing cache slot was always
inconclusive since the protocol does not specify a dedicated error pdu
that could be used to detect the failure via ExpectRdpegfxPdu.

The obvious expected behaviour (as implemented by mstsc) is to close the
egfx dynamic channel and this can be tested now.