Merged PR 1722: [Kerberos] Remove the case involved with krbtgt account.
Remove the case **Krbtgt_DES_PreAuthentication_Fail** which is involved with **krbtgt** account.

According to the doc below, the **krbtgt** account is disabled by default and cannot be enabled in Active Directory. The scenario of the case is not reasonable, since it is not possible to make it pass in current OSes nowadays.

https://docs.microsoft.com/en-us/windows/security/identity-protection/access-control/active-directory-accounts#sec-krbtgt

The test design for this case cannot be found in **Kerberos_ServerTestDesignSpecification.md** either.