Merged PR 1733: [Kerberos] Test the no PAC scenario with another service that supports decryption.
Test the no PAC scenario with another service that supports decryption.
To grant the access to the LADP Service in the local realm will get an undecryptable ticket returned due to the change of KDC's password.
According to the test design specification, it is not a specific test case to test the LDAP Service functionalities.
The account under test is replaced with another user account which is associated with a service instead of LDAP Service account.