Merged PR 1836: [RDPClient] Split the S2_AutoDetect_RTTMeasure case into 2 separated reliable and lossy cases.
DTLS randomly failed to decrypt the messages in the original case when a lossy RDPEUDP connection is established after the establishment of a reliable RDPEUDP connection.

According to the test scenario and test case design, there is no need to test the client behavior following the message flow in the original case. And the test case failure disappeared after split the case by RDPEDUP connection types.