Merged PR 7981351: Updated build file for RDP Server plugin to include files required for the plugin
Fixes an issue with the build file for the RDP Server plugin.

The output generated from the build for the plugin currently does not include the config file and the docs, both of which are required by the PTM Service installation flow.

This PR fixes the issue by including the files in the build.

## Pull Request Checklist
### General
- [ ] Are all regression test passed?
- [ ] Are there any test cases that will expose unfixed TDIs or Windows bugs?

### New Test Case
- [ ] Have Design Spec and User Guide been updated?
- [ ] Can all the test cases be loaded and executed by PTM & PTMCli?
- [ ] Can the related changes support multiple platform(Windows, Linux, MacOS)?

### SDK Changes
- [ ] Are all related test suites Regression passed?

Updated build file for server plugin to include files required for installing plugins in PTM service