Merged PR 8585692: Fix to resolve bug with RDPEUDP socket connection on multiple auto-detect attempts.
This PR adds a fix for a bug that is triggered on PTM Service when a user tries to auto-detect more than once in the same session for the RDP client plugin.

The fix closes the RDPEUDP socket connection after each attempt.

## Pull Request Checklist
### General
- [ ] Are all regression test passed?
- [ ] Are there any test cases that will expose unfixed TDIs or Windows bugs?

### New Test Case
- [ ] Have Design Spec and User Guide been updated?
- [ ] Can all the test cases be loaded and executed by PTM & PTMCli?
- [ ] Can the related changes support multiple platform(Windows, Linux, MacOS)?

### SDK Changes
- [ ] Are all related test suites Regression passed?

Added fix to resolve bug with RDPEUDP socket connection when auto-detect is run more than once for RDP client